Rename some JS folders and update references
* block -> blocks
* component -> components

This makes them more consistent with naming in Gutenberg.
