+++ public_html/wp-content/mu-plugins/blocks/source/blocks.js
@@ -6,7 +6,7 @@ const { registerBlockType } = wp.blocks;
 /**
  * Internal dependencies
  */
-import { BLOCKS } from './block';
+import { BLOCKS } from './blocks/'; // Trailing slash required to differentiate the folder from the file.
 
 BLOCKS.forEach( ( { NAME, SETTINGS } ) => {
 	registerBlockType( NAME, SETTINGS );
