+++ public_html/wp-content/mu-plugins/blocks/source/blocks/organizers/block-controls.js
@@ -13,8 +13,8 @@ const { __ }                  = wp.i18n;
 /**
  * Internal dependencies
  */
-import { PlaceholderSpecificMode } from '../../component/block-controls';
-import { getOptionLabel }          from '../../component/item-select';
+import { PlaceholderSpecificMode } from '../../components/block-controls';
+import { getOptionLabel }          from '../../components/item-select';
 import { BlockContent }            from './block-content';
 import { Select }                  from './select';
 import { LABEL }                   from './index';
@@ -42,8 +42,8 @@ export class BlockControls extends Component {
 				);
 				break;
 
-			case 'wcb_speaker' :
-			case 'wcb_speaker_group' :
+			case 'wcb_organizer' :
+			case 'wcb_organizer_team' :
 				output = (
 					<PlaceholderSpecificMode
 						label={ getOptionLabel( mode, options.mode ) }
@@ -79,8 +79,7 @@ export class BlockControls extends Component {
 
 						<div className="wordcamp-block-edit-mode-option">
 							<Select
-								icon={ icon }
-								label={ __( 'Choose specific speakers or groups', 'wordcamporg' ) }
+								label={ __( 'Choose specific organizers or teams', 'wordcamporg' ) }
 								{ ...this.props }
 							/>
 						</div>
