+++ public_html/wp-content/mu-plugins/blocks/source/blocks/sessions/index.js
