+++ public_html/wp-content/mu-plugins/blocks/source/components/block-content/index.js
