+++ public_html/wp-content/mu-plugins/blocks/source/components/image/featured-image.js
