+++ public_html/wp-content/mu-plugins/blocks/source/components/image/inspector-controls.js
